Fix Bug with Attached Device Not Detected
The signature is used to determin if the table should refresh.
The field the signature is based off of is nil for some phones.
Modify the signature to include the number of devices. This
will now only cause issues if a phone has been swaped and
both phones do not have a first boot field.

Additional Changes
* simplied device height calculation
* removed dead code